% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_draws.R
\docType{data}
\name{posterior_draws}
\alias{posterior_draws}
\title{Posterior draws of concentration parameters}
\format{
A tibble with 4000 rows (posterior) draws and 2 columns (models).
}
\usage{
posterior_draws
}
\description{
A tibble that contains posterior draws of the estimated concentration
parameters, based on the models presented in the main text of Cohen and
Hanretty (2022).
}
\keyword{datasets}
