% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharesimulatoR-package.R
\docType{package}
\name{sharesimulatoR-package}
\alias{sharesimulatoR}
\alias{sharesimulatoR-package}
\title{Data and Methods for Analyzing Comparative Vote Switching Data}
\description{
The \code{sharesimulatorR} package offers functions for
simulating realistic party systems of a given size.
}
\details{
The 'sharesimulatoR' package.
}
\references{
\enumerate{
  \item Cohen, Denis and Hanretty, Chris (2022). Simulating Party Shares.
   Unpublished manuscript.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrishanretty/sharesimulatoR}
  \item Report bugs at \url{https://github.com/chrishanretty/sharesimulatoR/issues}
}

}
\author{
\strong{Maintainer}: Chris Hanretty \email{chris.hanretty@rhul.ac.uk}

Authors:
\itemize{
  \item Denis Cohen \email{denis.cohen@uni-mannheim.de}
}

}
\keyword{internal}
