% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_sharesimulatoR.R
\name{shiny_sharesimulatoR}
\alias{shiny_sharesimulatoR}
\title{Interactively simulate realistic party systems in a Shiny GUI}
\usage{
shiny_sharesimulatoR()
}
\description{
Runs a ShinyApp that guides users through the steps
required for simulate party shares of a given size.
}
