% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shares.R
\name{simulate_shares}
\alias{simulate_shares}
\title{Simulate party shares for a party system of a given size}
\usage{
simulate_shares(
  N0,
  what = c("seat-winning parties", "vote-winning parties"),
  basis = c("Posterior draws of alpha", "Freely chosen deterministic alpha",
    "Freely chosen stochastic alpha"),
  alpha = NULL,
  sd_alpha = 0,
  n_sim = 1000L,
  seed = 20220920L
)
}
\arguments{
\item{N0}{Number of seat-winning/vote-winning parties (i.e., party system
size).}

\item{what}{What to simulate: seat or vote-winning parties. Affects the
choice of the concentration parameter if
\code{basis == "Posterior draws of alpha"}.}

\item{basis}{Specifies which concentration parameter will be used to
simulate party systems. \code{"Posterior draws of alpha"} uses 4000
posterior draws from the best-performing models from Cohen and Hanretty
(2022). \code{"Freely chosen deterministic alpha"} generates \code{n_sim}
draws using a fixed concentration parameter supplied per \code{alpha}.
\code{"Freely chosen sthochastic alpha"} generates \code{n_sim}
draws of alpha from a normal distribution with mean \code{alpha} and
standard deviation \code{sd_alpha} and uses each draw for one
corresponding party system simulation.}

\item{alpha}{User-supplied (mean) concentration parameter.}

\item{sd_alpha}{User-supplied standard deviation of the concentration
parameter.}

\item{n_sim}{Number of simulations. Fixed to the number of posterior draws
supplied by the package data unless users request freely chosen
deterministic/stochastic concentration parameters."}

\item{seed}{A seed set of reproducability.}
}
\value{
Returns a matrix which contains \code{n_sim} simulations (in rows) for N0 parties (in columns).
}
\description{
Simulates vote or seat shares for a party system of size N0, given the supplied concentration parameter.
}
